requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
python-telegram-bot>=20.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.2.0
